# swagger.yaml
openapi: 3.0.0
info:
  title: Car Contract Management System API
  description: 중고차 계약 관리 시스템 API 문서
  version: 1.0.0
  contact:
    name: API Support
    email: support@carmate.com
servers:
  - url: http://localhost:${process.env.PORT || 3000}
    description: Development server
    # 배포 전이므로 주석 처리
  # - url: ${process.env.PROD_SERVER_URL || 'https://api.carmate.com'}
  #  description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /cars:
    post:
      tags:
        - Cars
      summary: 차량 등록
      description: 새로운 차량 정보 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCarRequest'
      responses:
        '201':
          description: 차량 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags:
        - Cars
      summary: 차량 목록 조회
      description: 등록된 차량 목록 조회
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [possession, contractProceeding, contractCompleted]
        - in: query
          name: searchBy
          schema:
            type: string
            enum: [carNumber, model]
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: 차량 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cars/models:
    get:
      tags:
        - Cars
      summary: 차량 모델 목록 조회
      description: 제조사별 차량 모델 목록 조회
      responses:
        '200':
          description: 차량 모델 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        manufacturer:
                          type: string
                        model:
                          type: array
                          items:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cars/{carId}:
    get:
      tags:
        - Cars
      summary: 차량 상세 조회
      description: 특정 차량의 상세 정보 조회
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 차량 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Cars
      summary: 차량 정보 수정
      description: 차량 정보 수정
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCarRequest'
      responses:
        '200':
          description: 차량 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Cars
      summary: 차량 삭제
      description: 차량 정보 삭제
      parameters:
        - in: path
          name: carId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 차량 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '차량 삭제 성공'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cars/upload:
    post:
      tags:
        - Cars
      summary: 차량 대용량 업로드
      description: CSV 파일을 통한 차량 정보 대용량 등록
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 차량 업로드 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'CSV 파일 처리가 완료되었습니다.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
      description: 페이지 번호
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        default: 8
      description: 페이지당 아이템 수
    KeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: 검색 키워드
  schemas:
...
    CreateCarRequest:
      type: object
      required:
        - carNumber
        - manufacturer
        - model
        - type
        - manufacturingYear
        - mileage
        - price
      properties:
        carNumber:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        type:
          $ref: '#/components/schemas/CarType'
        manufacturingYear:
          type: integer
        mileage:
          type: integer
        price:
          type: number
        accidentCount:
          type: integer
          default: 0
        explanation:
          type: string
        accidentDetails:
          type: string
    UpdateCarRequest:
      type: object
      properties:
        carNumber:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        manufacturingYear:
          type: integer
        mileage:
          type: integer
        price:
          type: number
        accidentCount:
          type: integer
        explanation:
          type: string
        accidentDetails:
          type: string
    CarResponse:
      type: object
      properties:
        id:
          type: integer
        carNumber:
          type: string
        manufacturer:
          type: string
        model:
          type: string
        type:
          $ref: '#/components/schemas/CarType'
        manufacturingYear:
          type: integer
        mileage:
          type: integer
        price:
          type: number
        accidentCount:
          type: integer
        explanation:
          type: string
        accidentDetails:
          type: string
        status:
          $ref: '#/components/schemas/CarStatus'
    CarListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItemCount:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/CarResponse'

    # Customer schemas
...
    CarType:
      type: string
      enum:
        - 경·소형
        - 준중·중형
        - 대형
        - 스포츠카
        - SUV
    CarStatus:
      type: string
      enum:
        - possession
        - contractProceeding
        - contractCompleted
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '잘못된 요청입니다'
            code: 'BAD_REQUEST'
            timestamp: '2024-01-01T00:00:00Z'
    Unauthorized:
      description: 인증이 필요합니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '로그인이 필요합니다'
            code: 'UNAUTHORIZED'
            timestamp: '2024-01-01T00:00:00Z'
    Forbidden:
      description: 권한이 없습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '접근 권한이 없습니다'
            code: 'FORBIDDEN'
            timestamp: '2024-01-01T00:00:00Z'
    NotFound:
      description: 리소스를 찾을 수 없습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '존재하지 않는 리소스입니다'
            code: 'NOT_FOUND'
            timestamp: '2024-01-01T00:00:00Z'
    Conflict:
      description: 충돌이 발생했습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '이미 존재하는 데이터입니다'
            code: 'CONFLICT'
            timestamp: '2024-01-01T00:00:00Z'